requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
opencv-python-headless
numpy
supabase
openai-whisper
ffmpeg-python
python-dotenv